changed scoring command groups
